Move non-coroutine tests to CodegenCXX
